Refactor GenerationController to function and enhance code quality
- Convert GenerationController class to generate_invoice_files() function
- Rename generation_controller.py -> invoice_generator.py for clarity
- Fix inline imports by moving to top-level in test and application files
- Resolve pylint variable naming conflicts using pytest fixture pattern
- Add .pylintrc configuration for maintainable linting standards
- Enable strict pylint checking for all categories (C,R,W,E,F) including tests
- Maintain 10.00/10 pylint score with enhanced quality standards
