+++ .pylintrc
@@ -1,0 +1,13 @@
+[MAIN]
+# Fail on any of these message categories
+fail-on = C,R,W,E,F
+
+# Files or directories to ignore
+ignore = playwright.config.py
+
+[DESIGN]
+# Maximum number of attributes for a class
+max-attributes = 8
+
+# Minimum number of public methods for a class  
+min-public-methods = 2
