+++ application/app.py
@@ -5,6 +5,8 @@ import os
 
 from flask import Flask
 
+from . import db, routes
+
 
 def create_app() -> Flask:
     """Create and configure Flask application."""
@@ -18,13 +20,9 @@ def create_app() -> Flask:
     configure_logging(app)
 
     # Initialize database
-    from . import db
-
     db.init_app(app)
 
     # Register routes
-    from . import routes
-
     app.register_blueprint(routes.bp)
 
     return app
@@ -39,5 +37,5 @@ def configure_logging(app: Flask) -> None:
 
 
 if __name__ == "__main__":
-    app = create_app()
-    app.run(host="0.0.0.0", port=5000, debug=True)
+    flask_app = create_app()
+    flask_app.run(host="0.0.0.0", port=5000, debug=True)
